api/pyproject.toml
[tool.poetry]
name = "auto-blog"
version = "0.1.0"
description = ""
authors = ["m.kumeta <m.kumeta@nakamu-farm.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
python-wordpress-xmlrpc = "^2.3"
openai = "^0.27.4"
python-dotenv = "^1.0.0"
langchain = "^0.0.150"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
autoflake = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
